Added DeferredNode
A DeferredNode is a node that only contains a callback function that is
called as soon as _walk_value of the node is called. The callback function
must return an instance of \Structr\Tree\Base\Node, which will then
called. This can be used for example for definitions that you need for
some calls but not for others that need to create an enum from a database
for example and you don't want to hit the database for each and every request.